pyproject.toml
[project]
name = "fitness"
version = "0.1.0"
description = "Analysis of my fitness data"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "psycopg[binary]>=3.3.2",
    "alembic>=1.18.0",
    "typing_extensions>=4.0.0",
    "PyJWT[crypto]>=2.10.1",
    "cryptography>=46.0.3",
]

[project.scripts]
app = "fitness.app:app"

[dependency-groups]
dev = [
    "coverage>=7.13.1",
    "ipython>=9.9.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.11",
    "ty>=0.0.11",
    "testcontainers[postgres]>=4.13.3",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = [
    "e2e: end-to-end API + DB workflow tests",
]
addopts = "--cov=fitness --cov-report html"
